requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.0.0
apache-airflow-providers-common-sql==1.8.0
apispec==6.3.0
arandr==0.1.10
asgiref==3.7.2
astroid==2.5.1
asttokens==2.0.4
async-timeout==4.0.3
attrs==23.1.0
automationhat==0.2.0
Babel==2.13.1
backoff==1.10.0
beautifulsoup4==4.9.3
blinker==1.4
blinkt==0.1.2
buttonshim==0.0.2
cachelib==0.9.0
Cap1xxx==0.1.3
certifi==2023.7.22
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
clickclick==20.10.2
colorama==0.4.6
colorzero==1.1
connexion==2.14.2
cryptography==3.3.2
cupshelpers==1.0
dbus-python==1.2.16
Deprecated==1.2.14
distro==1.8.0
dnspython==2.4.2
docutils==0.20.1
drumhat==0.1.0
email-validator==1.3.1
envirophat==1.0.0
exceptiongroup==1.1.3
ExplorerHAT==0.4.2
Flask==2.2.5
Flask-Babel==2.0.0
Flask-JWT-Extended==4.5.3
Flask-Limiter==3.5.0
Flask-Login==0.6.3
Flask-SQLAlchemy==2.5.1
flask-wtf==1.2.1
fourletterphat==0.1.0
frozenlist==1.4.0
googleapis-common-protos==1.61.0
gpiozero==1.6.2
graphviz==0.20.1
greenlet==3.0.1
grpcio==1.59.2
gunicorn==21.2.0
h11==0.14.0
html5lib==1.1
httpcore==0.16.3
httpx==0.23.3
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
inflection==0.5.1
isort==5.6.4
itsdangerous==2.1.2
jedi==0.18.0
Jinja2==3.1.2
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
lazy-object-proxy==1.9.0
limits==3.6.0
linkify-it-py==2.0.2
lockfile==0.12.2
logilab-common==1.8.1
lxml==4.6.3
Mako==1.2.4
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
mccabe==0.6.1
mdit-py-plugins==0.4.0
mdurl==0.1.2
microdotphat==0.2.1
mote==0.0.4
motephat==0.0.3
multidict==6.0.4
mypy==0.812
mypy-extensions==0.4.3
numpy==1.26.2
oauthlib==3.1.0
openai==1.3.8
opentelemetry-api==1.20.0
opentelemetry-exporter-otlp==1.20.0
opentelemetry-exporter-otlp-proto-common==1.20.0
opentelemetry-exporter-otlp-proto-grpc==1.20.0
opentelemetry-exporter-otlp-proto-http==1.20.0
opentelemetry-proto==1.20.0
opentelemetry-sdk==1.20.0
opentelemetry-semantic-conventions==0.41b0
ordered-set==4.1.0
packaging==23.2
pandas==2.1.4
pantilthat==0.0.7
parso==0.8.1
pathspec==0.11.2
pendulum==2.1.2
pexpect==4.8.0
pgzero==1.2
phatbeat==0.1.1
pianohat==0.1.0
picamera==1.13
picamera2==0.3.12
pidng==4.0.9
piexif==1.1.3
piglow==1.2.5
pigpio==1.78
Pillow==8.1.2
pluggy==1.3.0
prison==0.2.1
protobuf==4.24.4
psutil==5.9.6
pycairo==1.16.2
pycparser==2.21
pycups==2.0.1
pydantic==2.4.2
pydantic-core==2.10.1
pygame==1.9.6
Pygments==2.16.1
PyGObject==3.38.0
pyinotify==0.9.6
PyJWT==2.8.0
pylint==2.7.2
PyOpenGL==3.1.5
pyOpenSSL==20.0.1
PyQt5==5.15.2
PyQt5-sip==12.8.1
pyserial==3.5b0
pysmbc==1.0.23
python-apt==2.2.1
python-daemon==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-nvd3==0.15.0
python-prctl==1.7
python-slugify==8.0.1
pytz==2023.3.post1
pytzdata==2020.1
PyYAML==6.0.1
rainbowhat==0.1.0
referencing==0.30.2
reportlab==3.5.59
requests==2.31.0
requests-oauthlib==1.0.0
requests-toolbelt==1.0.0
responses==0.12.1
rfc3339-validator==0.1.4
rfc3986==1.5.0
rich==13.6.0
rich-argparse==1.4.0
roman==2.0.0
rpds-py==0.10.6
RPi.GPIO==0.7.0
RTIMULib==7.2.1
scrollphat==0.0.7
scrollphathd==1.2.1
Send2Trash==1.6.0b1
sense-hat==2.4.0
setproctitle==1.3.3
simplejpeg==1.6.4
simplejson==3.17.2
six==1.16.0
skywriter==0.0.7
sn3218==1.2.7
sniffio==1.3.0
soupsieve==2.2.1
spidev==3.5
SQLAlchemy==1.4.50
SQLAlchemy-JSONField==1.0.1.post0
SQLAlchemy-Utils==0.41.1
sqlparse==0.4.4
ssh-import-id==5.10
tabulate==0.9.0
tenacity==8.2.3
termcolor==2.3.0
text-unidecode==1.3
thonny==4.0.1
toml==0.10.1
touchphat==0.0.1
tqdm==4.66.1
twython==3.8.2
typed-ast==1.4.2
typing-extensions==4.8.0
tzdata==2023.3
uc-micro-py==1.0.2
unicodecsv==0.14.1
unicornhathd==0.0.4
urllib3==1.26.18
v4l2-python3==0.3.2
webencodings==0.5.1
Werkzeug==2.2.3
wrapt==1.15.0
WTForms==3.0.1
yarl==1.9.2
zipp==3.17.0
